Update README.md
Step 1: Write a PowerShell script to create a local security group within the Windows VM from assignment 8a.  The script should also add the created non-admin user to that group.

Step 2: Add a line to the script that will create a folder c:\AssignmentB, and give the created security group full permission to that folder.

 

To complete the assignment, submit this script for grading.